float support

+++ app/Http/Controllers/CalcController.php
@@ -7,10 +7,17 @@ use Illuminate\Http\Request;
 class CalcController extends Controller
 {
     //
-    public function div(int $a, int $b)
+    public function div(float $a, float $b)
     {
         return response()->json(
             ["result" => $a / $b]
         );
     }
+
+    // public function mul(int $a, int $b)
+    // {
+    //     return response()->json(
+    //         ["result" => $a * $b]
+    //     );
+    // }
 }
